features/cms/supabase/flask-implementation/pyproject.toml
[project]
name = "flask-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "supabase>=2.27.2",
]
